Resolving the review comments
